web/requirements.txt
Django==1.8.3
psycopg2
Pillow
djangorestframework==3.1